+++ Nonstick Summer 2025/Assets/Card Templates/Charming Observation.asset
@@ -15,4 +15,5 @@ MonoBehaviour:
   _energyCost: -2
   _emotion: 1
   _intention: 2
-  _stamps: []
+  _stamps:
+  - {fileID: 11400000, guid: 7f02ae394bcf2374b83edb05d6f6bd4d, type: 2}
