+++ Nonstick Summer 2025/Assets/Modifier Pickups/Energy Bonus.asset
@@ -10,8 +10,8 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 305ed51a54b892448a18e85c1f8ba3a0, type: 3}
-  m_Name: Extrovert
+  m_Name: Energy Bonus
   m_EditorClassIdentifier: 
   MaxCardsApplied: 1
   MinCardsApplied: 1
-  StampToApply: {fileID: 11400000, guid: b8e51bfdf95cfa14c859212c7f4f3ec9, type: 2}
+  StampToApply: {fileID: 11400000, guid: 7f02ae394bcf2374b83edb05d6f6bd4d, type: 2}
