+++ Nonstick Summer 2025/Assets/Modifier Pickups/Energy Bonus.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b8e51bfdf95cfa14c859212c7f4f3ec9
+guid: f2ca2cad2d5f3834187f980423098bc9
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
