+++ Nonstick Summer 2025/Assets/Scenes/Firstplayablelevel.unity
@@ -54784,6 +54784,10 @@ PrefabInstance:
       propertyPath: m_Name
       value: GameManager
       objectReference: {fileID: 0}
+    - target: {fileID: 8412335191130524943, guid: 1b5ab685e734f574c9d1e7815a237b25, type: 3}
+      propertyPath: 'startingModifiers.Array.data[1]'
+      value: 
+      objectReference: {fileID: 11400000, guid: f2ca2cad2d5f3834187f980423098bc9, type: 2}
     m_RemovedComponents: []
     m_RemovedGameObjects: []
     m_AddedGameObjects:
@@ -71450,7 +71454,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 681796896327735944, guid: 95bcd8e9c8d5eb549b0d0d21529b58db, type: 3}
       propertyPath: m_LocalPosition.z
-      value: -8.5
+      value: 12.82
       objectReference: {fileID: 0}
     - target: {fileID: 681796896327735944, guid: 95bcd8e9c8d5eb549b0d0d21529b58db, type: 3}
       propertyPath: m_LocalRotation.w
