+++ Nonstick Summer 2025/Assets/Scripts/Cards/Stamps/EnergyBonusStamp.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 4af2b4767ce194b4ea5fab8f04a0108a
