+++ Nonstick Summer 2025/Assets/Scripts/Cards/Stamps/System/StatOperator.cs
@@ -1,0 +1,5 @@
+public enum StatOperator
+{
+    Add,
+    Multiply,
+}
