+++ Nonstick Summer 2025/Assets/Scripts/Cards/Stamps/System/StatOperator.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 77b30a9027098bb419e57ce2a514e7c4
