+++ Nonstick Summer 2025/Assets/Stamps/Energy Bonus.asset
@@ -9,13 +9,11 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: d97fda75aca9b97438e9d568896316cf, type: 3}
-  m_Name: Energy Cost Adder
+  m_Script: {fileID: 11500000, guid: 4af2b4767ce194b4ea5fab8f04a0108a, type: 3}
+  m_Name: Energy Bonus
   m_EditorClassIdentifier: 
   TriggerCondition: 0
   Icon: {fileID: 21300000, guid: 4916a53627bd9d3479a03c9aedb8ef1f, type: 3}
-  StampName: Extroverted
-  ShortDescription: +N Energy for free
-  stat: 0
-  mode: 0
-  value: 3
+  StampName: Energy Bonus
+  ShortDescription: Gives +3 energy after card is played
+  BonusEnergy: 3
