+++ Nonstick Summer 2025/Assets/Stamps/Energy Bonus.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7ea86a966298e524e81d7076ef263b99
+guid: 7f02ae394bcf2374b83edb05d6f6bd4d
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
