+++ Nonstick Summer 2025/Assets/Stamps/Relationship Multiplier.asset
@@ -15,7 +15,6 @@ MonoBehaviour:
   TriggerCondition: 0
   Icon: {fileID: 21300000, guid: f21bc02bf72f7c141aa8a92e4643a093, type: 3}
   StampName: 2x Relationship
-  ShortDescription: 
-  stat: 1
+  ShortDescription: Multiplies change in relationship by 2
   mode: 1
   value: 2
